# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Car:
    properties:
      brand:
        type: string
      comment:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      kilometers:
        type: integer
      model:
        type: string
      ownerID:
        description: Engine       Engine
        type: integer
      owners_number:
        type: integer
      placement:
        type: string
      price:
        type: integer
      status:
        type: string
      updatedAt:
        type: string
      vin_code:
        type: string
      year:
        type: integer
    type: object
  models.Engine:
    properties:
      cilinders:
        type: integer
      consumption:
        type: number
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      fuel:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.EngineRequest:
    properties:
      cilinders:
        type: integer
      consumption:
        type: number
      fuel:
        type: string
      name:
        type: string
    type: object
  models.Error:
    properties:
      error:
        type: string
    type: object
  models.Message:
    properties:
      message:
        type: string
    type: object
  models.PaymentIntent:
    properties:
      _id:
        type: string
      amount:
        type: number
      canceled:
        type: integer
      createdAt:
        type: string
      currency:
        type: string
      customer:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      secret:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  models.PaymentIntentList:
    properties:
      data:
        items:
          $ref: '#/definitions/models.PaymentIntent'
        type: array
      length:
        type: integer
    type: object
  models.RequestLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.RequestRegistration:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.VINRequest:
    properties:
      vin_code:
        type: string
    required:
    - vin_code
    type: object
  models.VINResponse:
    properties:
      country:
        type: string
      manufacturer:
        type: string
      region:
        type: string
      vds:
        type: string
      vin:
        type: string
      vis:
        type: string
      wmi:
        type: string
      years:
        items:
          type: integer
        type: array
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is documentation for Cars Sales App API for all user operations
  title: Cars Sales App API
  version: "0.6"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login to app
      parameters:
      - description: Email, Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Login
      tags:
      - Authorization
  /auth/registration:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Name, Email, Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Registration
      tags:
      - Authorization
  /cars/all:
    get:
      consumes:
      - application/json
      description: Get all cars
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Car'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Cars
      tags:
      - Cars
  /cars/create:
    post:
      consumes:
      - application/json
      description: Create a car for sale
      parameters:
      - description: Car
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Car'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
      summary: Cars
      tags:
      - Cars
  /cars/my:
    get:
      consumes:
      - application/json
      description: Get owned Cars
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Car'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
      summary: Cars
      tags:
      - Cars
  /engine/all:
    get:
      consumes:
      - application/json
      description: Get All engines
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Engine'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Engine CRUD
      tags:
      - Engine
  /engine/create:
    post:
      consumes:
      - application/json
      description: Create new Engine
      parameters:
      - description: Engine Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.EngineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Engine'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Engine CRUD
      tags:
      - Engine
  /engine/delete/:id:
    delete:
      consumes:
      - application/json
      description: Delete engine by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Engine CRUD
      tags:
      - Engine
  /engine/update/:id:
    put:
      consumes:
      - application/json
      description: Update engine info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Engine'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Engine CRUD
      tags:
      - Engine
  /paymnet_intent/:id:
    get:
      consumes:
      - application/json
      description: Get Payment Intent by ID
      parameters:
      - description: Payment Intent ID
        in: path
        name: payment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentIntent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Payment Intent Operation
      tags:
      - Payment Intent
  /paymnet_intent/all:
    get:
      consumes:
      - application/json
      description: Get Customers Payment Intents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PaymentIntentList'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
      summary: Payment Intent Operation
      tags:
      - Payment Intent
  /paymnet_intent/cancel:
    post:
      consumes:
      - application/json
      description: Cancel Payment Intent
      parameters:
      - description: id
        in: body
        name: request
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Payment Intent Operation
      tags:
      - Payment Intent
  /paymnet_intent/create:
    post:
      consumes:
      - application/json
      description: Create Payment Intent
      parameters:
      - description: Email, Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PaymentIntent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
      summary: Payment Intent Operation
      tags:
      - Payment Intent
  /vincode/check:
    post:
      consumes:
      - application/json
      description: Use VIN-code for more details
      parameters:
      - description: VIN-code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VINRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VINResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: VIN
      tags:
      - VIN
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"